Added pundit task policy

+++ app/controllers/tasks_controller.rb
@@ -1,54 +1,64 @@
 # frozen_string_literal: true
 
 class TasksController < ApplicationController
-  before_action :authenticate_user_using_x_auth_token, except: [:new, :edit]
-
+  after_action :verify_authorized, except: :index
+  after_action :verify_policy_scoped, only: :index
+  before_action :authenticate_user_using_x_auth_token
   before_action :load_task, only: %i[show update destroy]
 
-  def show
-    render status: :ok, json: { task: @task }
-  end
   def index
-    tasks = Task.all
-    render status: :ok, json: { tasks: tasks }
+    #------new line added here------
+    @tasks = policy_scope(Task)
+    #-----end of added line-------
   end
 
   def create
     @task = Task.new(task_params.merge(creator_id: @current_user.id))
+    authorize @task
     if @task.save
-      render status: :ok, json: { notice: t("successfully_created", entity: "Task") }
+      render status: :ok,
+             json: { notice: t("successfully_created", entity: "Task") }
     else
       errors = @task.errors.full_messages.to_sentence
       render status: :unprocessable_entity, json: { errors: errors }
     end
   end
 
-  def task_params
-    params.require(:task).permit(:title, :user_id)
+  def show
+    authorize @task
+    @task_creator = User.find(@task.creator_id).name
   end
 
   def update
+    authorize @task
     if @task.update(task_params)
-      render status: :ok, json: { notice: "Successfully updated task." }
+      render status: :ok, json: {}
     else
-      render status: :unprocessable_entity, json: { errors: @task.errors.full_messages.to_sentence }
+      render status: :unprocessable_entity,
+             json: { errors: @task.errors.full_messages.to_sentence }
     end
   end
 
   def destroy
+    authorize @task
     if @task.destroy
-      render status: :ok, json: { notice: "Successfully deleted task." }
+      render status: :ok, json: {}
     else
-      render status: :unprocessable_entity, json: { errors: @task.errors.full_messages.to_sentence }
+      render status: :unprocessable_entity,
+             json: { errors: @task.errors.full_messages.to_sentence }
     end
   end
 
   private
 
+    def task_params
+      params.require(:task).permit(:title, :user_id)
+    end
+
     def load_task
       @task = Task.find_by_slug!(params[:slug])
-    rescue ActiveRecord::RecordNotFound => errors
-      render json: { errors: errors }
+    rescue ActiveRecord::RecordNotFound => e
+      render json: { errors: e }, status: :not_found
     end
 end
